+++ simulator-ui/scripts/w1.py
@@ -22,7 +22,7 @@ NOT 		= "org.hanns.logic.crisp.gates.impl.NOT";
 def or_node(name):
 	g = NodeGroup("OR", True);
 	g.addNode(OR, "OR", "java");
-	module = NeuralModule(name+'_OR', g)
+	module = NeuralModuleMI(name+'_OR', g)
 	module.createEncoder("logic/gates/ina", "bool", 1)
 	module.createEncoder("logic/gates/inb", "bool", 1)
 	module.createDecoder("logic/gates/outa", "bool", 1)
@@ -48,6 +48,9 @@ net.add(myNOT)
 # Wire inputs
 net.connect(gen1, myOR.getTermination('logic/gates/ina'))
 net.connect(gen2, myOR.getTermination('logic/gates/inb'))
+
+# TODO
+#net.connect(gen2, myOR.aWTT('logic/gates/inb', 0.1))
 
 net.connect(myOR.getOrigin('logic/gates/outa'), myNOT.getTermination('logic/gates/ina'), weight=1)